[change] some changes in catch blocks

+++ src/infrastructure/controllers/RecordController.ts
@@ -1,5 +1,6 @@
 import { RequestHandler } from 'express';
 import { RecordService } from '../../domain/services/RecordService';
+import {ApiError} from "../../errors/API.error";
 
 export class RecordController {
     constructor(private readonly _service: RecordService) {}
@@ -24,13 +25,18 @@ export class RecordController {
         const recordId = Number(req.params.record_id);
 
         try {
+            const record = await this._service.getById(recordId);
+
+            if (record.userId !== req.body._userId) {
+                return res.status(403).json('You cannot delete this');
+            }
             await this._service.delete(recordId);
 
             return res
                 .status(200)
                 .json({status: res.statusCode, message: 'Record has been deleted!'});
         } catch (err) {
-            if (err instanceof Error) {
+            if (err instanceof ApiError) {
                 return res
                     .status(404)
                     .json({status: res.statusCode, message: `${err.message}`});
@@ -44,7 +50,7 @@ export class RecordController {
 
             return res.status(201).json({ record: record });
         } catch (err) {
-            if (err instanceof Error) {
+            if (err instanceof ApiError) {
                 return res
                     .status(404)
                     .json({status: 404, message: `${ err.message }`});
@@ -70,8 +76,11 @@ export class RecordController {
 
             return res.status(200).json({records: records});
         } catch (err) {
-            console.log(err)
-            return res.status(404).send('Some error happened');
+            if (err instanceof ApiError) {
+                return res
+                    .status(404)
+                    .json({status: 404, message: `${ err.message }`});
+            }
         }
     };
 }
